+++ dl/src/main/scala/com/intel/analytics/bigdl/nn/Nms.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.bigdl.models.fasterrcnn.utils
+package com.intel.analytics.bigdl.nn
 
 import com.intel.analytics.bigdl.tensor.Tensor
